requirements.txt
Django==3.0.6
django_rest_swagger==2.2.0
djangorestframework==3.11.0
